simulator/requirements.txt
pandas
kafka-python
numpy